Task6

+++ Tyuiu.FilatovDK.Sprint5.Task6.V20.Lib/DataService.cs
@@ -3,7 +3,7 @@ namespace Tyuiu.FilatovDK.Sprint5.Task6.V20.Lib
 {
     public class DataService : ISprint5Task6V20
     {
-        public double LoadFromDataFile(string path)
+        public int LoadFromDataFile(string path)
         {
 
             int count = 0;
